+++ test-resources/examples/ex2/expected.clj
@@ -1,3 +1,5 @@
+(ns test)
+
 (defn analyze-data [data]
   (let [processed (map process-data data)
         filtered (filter relevant? processed)