+++ test-resources/examples/ex2/left.clj
@@ -1,3 +1,4 @@
+(ns test)
 (defn analyze-data [data]
   (let [processed (map process-item data)
         filtered (filter relevant? processed)