+++ test-resources/examples/ex2/right.clj
@@ -1,3 +1,4 @@
+(ns test)
 (defn analyze-data [data]
   (let [processed (map process-data data)
         results (reduce combine-results {} processed)]